+++ NEWS.md
@@ -21,7 +21,7 @@ We reworked the infrastructure around adding and removing tabs using `new_add_ta
 - Allow initialising a dock with no panels (default to `list()`).
 - Added `input[["<dock_ID>_initialized"]]` within an `onRender` callback. Allows to track when the dock is ready
 to perform actions server side.
-- In `add_panel()`: if no `referencePanel` or `referenceGroup` is provided, the panel is added relative to the [container](https://dockview.dev/docs/core/panels/add#relative-to-the-container).
+- In `add_panel()`: if no `referencePanel` or `referenceGroup` is provided, the panel is added relative to the [container](https://dockview.dev/docs/core/panels/add/#relative-to-the-container).
 - Fix: `get_groups_ids()` now correctly returns all group ids (nested groups were not returned).
 - Fix [#52](https://github.com/cynkra/dockViewR/issues/52): Reworked `add_tab` parameter in `dock_view()`. By default, there is a `default_add_tab_callback()` that sets `input[["<dock_ID>_panel-to-add"]]`, so you can create observers with custom logic, including removing the panel with `add_panel()`. An example of usage is available at <https://github.com/cynkra/dockViewR/blob/main/inst/examples/add_panel/app.R>.
 - Fix: typo in `abyss-spaced` theme caused the theme not to be applied.